requirements.txt
# Core ML and AI libraries
torch>=2.0.0
torchvision>=0.15.0
efficientnet-pytorch>=0.7.1

# Web application framework
streamlit>=1.28.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Data manipulation and analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pathlib
logging
datetime
typing

# Development and testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Optional: GPU acceleration
# torch-audio>=2.0.0  # Uncomment if needed
# torchtext>=0.15.0   # Uncomment if needed
